updated grammar. I think it's complete
"not" is not a comparison operator
added canMove (bool) and getCarPosition (number)
changed make to define
added dependence on newlines (or semicolons, for our debugging
purposes) (also helped resolve a shift/reduce conflict)
added function calls
we will have to do a lot of type verification in the semantic analysis.
according to the grammar, "hello" * myVar myFunc + 2 is a valid
expression.